Create WEWOCHRODB
In this example, we're using the @octokit/rest package to create a new Octokit client with an access token. We then define a getRepository function that takes an owner and repo parameter, and uses the repos.get, repos.getBranch, and git.getCommit methods to retrieve information about the repository and its latest commit. Finally, we print out some information about the repository and commit to the console.

Note that this is just a simple example, and there are many other ways that you can use the Git Database and Git Refs APIs to build Node.js tools that interact with GitHub repositories.