Create NewDBAuth
This code creates a new file named newfile.txt in the default branch of a GitHub repository, using the Git Database API and the Octokit library.

Note that this is just one example of how to use the Git Database API, and there are many other ways to use this API to interact with GitHub repositories. The specifics of how to use the API may vary depending on your use case, so it's important to refer to the GitHub API documentation for more information and examples.