Create Dashboard
In this example, we have a CompliancePolicy class that represents a workplace compliance policy, with a name and a boolean value indicating whether the policy is enforced. We also have a ComplianceDashboard class that contains a list of CompliancePolicy objects and a method to display them in a dashboard format.

In the main function, we create a ComplianceDashboard object and call the displayDashboard method to display the list of compliance policies and their enforcement status.

This compliance dashboard can be used by businesses to monitor their compliance with workplace regulations and policies, and to ensure that all employees are aware of and following these policies. The dashboard can also be customized to include additional policies or information as needed.