Create inputRANDOM
In this updated example, we've added two new inputs to the dashboard: one for searching the compliance databases and one for checking compliance violations. We've also implemented randomized triggering for the compliance check using `setInterval`. The `setInterval` function executes the `checkCompliance` function at a random interval between 1 hour and 24 hours (in milliseconds). If a compliance violation is detected, an alert or other action can be triggered.

It's important to note that while randomized triggering can be useful for testing the compliance system, compliance violations should always be addressed promptly and in a consistent manner. Randomized triggering should not be used to create an environment of distrust or to undermine the effectiveness of the compliance program.