Create syncMODS
In this module, we define a function called `syncWithSystem` that generates a random integer between 1 and 3 to select a system to synchronize with. We use a switch statement to select the appropriate system based on the random integer.

In each case, we would write the code to synchronize with the selected system. This code will vary depending on the system being synchronized with, but may involve making API requests or updating a shared database.

After synchronizing with the selected system, we log a message to the console indicating that the synchronization was successful.

Finally, we export the `syncWithSystem` function for use in other modules. This allows us to synchronize the website with a random current system at regular intervals, improving the accuracy and reliability of compliance data.